Add gen_*.py to create samples of torchscript
