+++ examples/load-torchscript/gen_resnet.py
@@ -1,3 +1,6 @@
+# Original source   https://pytorch.org/tutorials/advanced/cpp_export.html
+# The artifact of this model is in https://github.com/hasktorch/libtorch-binary-for-ci/releases/tag/1.4.0.
+
 import torch
 import torchvision
 
@@ -10,3 +13,6 @@ example = torch.rand(1, 3, 224, 224)
 # Use torch.jit.trace to generate a torch.jit.ScriptModule via tracing.
 traced_script_module = torch.jit.trace(model, example)
 traced_script_module.save("traced_resnet_model.pt")
+
+script_module = torch.jit.script(model)
+script_module.save("resnet_model.pt")
