+++ hasktorch/hasktorch.cabal
@@ -49,6 +49,7 @@ library
                     , Torch.Serialize
                     , Torch.Random
                     , Torch.Script
+                    , Torch.Utils
  hs-source-dirs:      src
  default-language:    Haskell2010
  ghc-options:         -fplugin GHC.TypeLits.Normalise -fplugin GHC.TypeLits.KnownNat.Solver -fplugin GHC.TypeLits.Extra.Solver -fconstraint-solver-iterations=0
@@ -66,6 +67,10 @@ library
                     , reflection
                     , singletons
                     , stm
+                    , JuicyPixels
+                    , vector
+                    , bytestring
+                    , safe-exceptions
 
 test-suite spec
   type:               exitcode-stdio-1.0
